requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
Jinja2==3.1.4
itsdangerous==2.2.0

gunicorn
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib

